+++ tests/test_cli.py
@@ -17,6 +17,8 @@ DEFAULT_FPS = 15
 DEFAULT_AF_SPEED = AF_SPEED_ENUM_NORMAL
 DEFAULT_AUTOFOCUS_MODE = AF_MODE_ENUM_CONTINUOUS
 DEFAULT_CONTROLS = []
+DEFAULT_TUNING_FILTER = None
+DEFAULT_TUNING_FILTER_DIR = None
 
 
 @pytest.fixture(autouse=True)
@@ -62,6 +64,18 @@ def test_split_resolution():
     assert height == 200
 
 
+def test_parse_tuning_filter():
+    from spyglass import cli
+    args = cli.get_args(['-tf', 'filter'])
+    assert args.tuning_filter == 'filter'
+
+
+def test_parse_tuning_filter_dir():
+    from spyglass import cli
+    args = cli.get_args(['-tfd', 'dir'])
+    assert args.tuning_filter_dir == 'dir'
+
+
 @patch("spyglass.server.run_server")
 @patch("spyglass.camera.init_camera")
 def test_init_camera_with_defaults(mock_spyglass_camera, mock_spyglass_server):
@@ -78,7 +92,9 @@ def test_init_camera_with_defaults(mock_spyglass_camera, mock_spyglass_server):
         DEFAULT_UPSIDE_DOWN,
         DEFAULT_FLIP_HORIZONTALLY,
         DEFAULT_FLIP_VERTICALLY,
-        DEFAULT_CONTROLS
+        DEFAULT_CONTROLS,
+        DEFAULT_TUNING_FILTER,
+        DEFAULT_TUNING_FILTER_DIR
     )
 
 
@@ -100,7 +116,9 @@ def test_init_camera_resolution(mock_spyglass_server, mock_spyglass_camera):
         DEFAULT_UPSIDE_DOWN,
         DEFAULT_FLIP_HORIZONTALLY,
         DEFAULT_FLIP_VERTICALLY,
-        DEFAULT_CONTROLS
+        DEFAULT_CONTROLS,
+        DEFAULT_TUNING_FILTER,
+        DEFAULT_TUNING_FILTER_DIR
     )
 
 
@@ -138,7 +156,9 @@ def test_init_camera_fps(mock_spyglass_server, mock_spyglass_camera):
         DEFAULT_UPSIDE_DOWN,
         DEFAULT_FLIP_HORIZONTALLY,
         DEFAULT_FLIP_VERTICALLY,
-        DEFAULT_CONTROLS
+        DEFAULT_CONTROLS,
+        DEFAULT_TUNING_FILTER,
+        DEFAULT_TUNING_FILTER_DIR
     )
 
 
@@ -160,7 +180,9 @@ def test_init_camera_af_manual(mock_spyglass_server, mock_spyglass_camera):
         DEFAULT_UPSIDE_DOWN,
         DEFAULT_FLIP_HORIZONTALLY,
         DEFAULT_FLIP_VERTICALLY,
-        DEFAULT_CONTROLS
+        DEFAULT_CONTROLS,
+        DEFAULT_TUNING_FILTER,
+        DEFAULT_TUNING_FILTER_DIR
     )
 
 
@@ -182,7 +204,9 @@ def test_init_camera_af_continuous(mock_spyglass_server, mock_spyglass_camera):
         DEFAULT_UPSIDE_DOWN,
         DEFAULT_FLIP_HORIZONTALLY,
         DEFAULT_FLIP_VERTICALLY,
-        DEFAULT_CONTROLS
+        DEFAULT_CONTROLS,
+        DEFAULT_TUNING_FILTER,
+        DEFAULT_TUNING_FILTER_DIR
     )
 
 
@@ -204,7 +228,9 @@ def test_init_camera_lens_position(mock_spyglass_server, mock_spyglass_camera):
         DEFAULT_UPSIDE_DOWN,
         DEFAULT_FLIP_HORIZONTALLY,
         DEFAULT_FLIP_VERTICALLY,
-        DEFAULT_CONTROLS
+        DEFAULT_CONTROLS,
+        DEFAULT_TUNING_FILTER,
+        DEFAULT_TUNING_FILTER_DIR
     )
 
 
@@ -226,7 +252,9 @@ def test_init_camera_af_speed_normal(mock_spyglass_server, mock_spyglass_camera)
         DEFAULT_UPSIDE_DOWN,
         DEFAULT_FLIP_HORIZONTALLY,
         DEFAULT_FLIP_VERTICALLY,
-        DEFAULT_CONTROLS
+        DEFAULT_CONTROLS,
+        DEFAULT_TUNING_FILTER,
+        DEFAULT_TUNING_FILTER_DIR
     )
 
 
@@ -248,7 +276,9 @@ def test_init_camera_af_speed_fast(mock_spyglass_server, mock_spyglass_camera):
         DEFAULT_UPSIDE_DOWN,
         DEFAULT_FLIP_HORIZONTALLY,
         DEFAULT_FLIP_VERTICALLY,
-        DEFAULT_CONTROLS
+        DEFAULT_CONTROLS,
+        DEFAULT_TUNING_FILTER,
+        DEFAULT_TUNING_FILTER_DIR
     )
 
 
@@ -270,7 +300,9 @@ def test_init_camera_upside_down(mock_spyglass_server, mock_spyglass_camera):
         True,
         DEFAULT_FLIP_HORIZONTALLY,
         DEFAULT_FLIP_VERTICALLY,
-        DEFAULT_CONTROLS
+        DEFAULT_CONTROLS,
+        DEFAULT_TUNING_FILTER,
+        DEFAULT_TUNING_FILTER_DIR
     )
 
 
@@ -292,7 +324,9 @@ def test_init_camera_flip_horizontal(mock_spyglass_server, mock_spyglass_camera)
         DEFAULT_UPSIDE_DOWN,
         True,
         DEFAULT_FLIP_VERTICALLY,
-        DEFAULT_CONTROLS
+        DEFAULT_CONTROLS,
+        DEFAULT_TUNING_FILTER,
+        DEFAULT_TUNING_FILTER_DIR
     )
 
 
@@ -314,7 +348,9 @@ def test_init_camera_flip_vertical(mock_spyglass_server, mock_spyglass_camera):
         DEFAULT_UPSIDE_DOWN,
         DEFAULT_FLIP_HORIZONTALLY,
         True,
-        DEFAULT_CONTROLS
+        DEFAULT_CONTROLS,
+        DEFAULT_TUNING_FILTER,
+        DEFAULT_TUNING_FILTER_DIR
     )
 
 
@@ -334,7 +370,9 @@ def test_init_camera_controls():
         DEFAULT_UPSIDE_DOWN,
         DEFAULT_FLIP_HORIZONTALLY,
         DEFAULT_FLIP_VERTICALLY,
-        ['brightness=-0.4','awbenable=false']
+        ['brightness=-0.4','awbenable=false'],
+        DEFAULT_TUNING_FILTER,
+        DEFAULT_TUNING_FILTER_DIR
     )
 
 
@@ -383,4 +421,52 @@ def test_run_server_with_orientation(mock_spyglass_server, mock_spyglass_camera,
         'streaming-url',
         'snapshot-url',
         expected_output
+    )
+
+
+@patch("spyglass.server.run_server")
+@patch("spyglass.camera.init_camera")
+def test_init_camera_using_only_tuning_filter_file(mock_spyglass_server, mock_spyglass_camera):
+    from spyglass import cli
+    import spyglass.camera
+    cli.main(args=[
+        '-tf', 'test',
+    ])
+    spyglass.camera.init_camera.assert_called_once_with(
+        DEFAULT_WIDTH,
+        DEFAULT_HEIGHT,
+        DEFAULT_FPS,
+        DEFAULT_AUTOFOCUS_MODE,
+        DEFAULT_LENS_POSITION,
+        DEFAULT_AF_SPEED,
+        DEFAULT_UPSIDE_DOWN,
+        DEFAULT_FLIP_HORIZONTALLY,
+        DEFAULT_FLIP_VERTICALLY,
+        DEFAULT_CONTROLS,
+        "test",
+        DEFAULT_TUNING_FILTER_DIR
+    )
+
+@patch("spyglass.server.run_server")
+@patch("spyglass.camera.init_camera")
+def test_init_camera_using_tuning_filters(mock_spyglass_server, mock_spyglass_camera):
+    from spyglass import cli
+    import spyglass.camera
+    cli.main(args=[
+        '-tf', 'test',
+        '-tfd', 'test-dir',
+    ])
+    spyglass.camera.init_camera.assert_called_once_with(
+        DEFAULT_WIDTH,
+        DEFAULT_HEIGHT,
+        DEFAULT_FPS,
+        DEFAULT_AUTOFOCUS_MODE,
+        DEFAULT_LENS_POSITION,
+        DEFAULT_AF_SPEED,
+        DEFAULT_UPSIDE_DOWN,
+        DEFAULT_FLIP_HORIZONTALLY,
+        DEFAULT_FLIP_VERTICALLY,
+        DEFAULT_CONTROLS,
+        "test",
+        "test-dir"
     )